Refactor calculator to work with a calendar internally
This provides for much more expressive calculations and allows for greater flexibility
within the class. Rather than having to construct an interval string, it is now possible
to perform arithmetic on specific fields which is much clearer and simpler in some cases.

This change also includes some initial tests for the calculator.